Pull common test-related code out of `core`
This allows `web` and `it` to depend on `core` and a couple of new, small subprojects, instead of `core`'s tests. This means slightly faster local builds and less hassle when things are in flux.

It will also allow other projects to make use of the Arbitrary instances and other utilities.