move to ExprOp
`AccumOp` and `Grouped` now handle any `EX`.

Everything else including all workflow-related types now uses the biggest coproduct, `ExprOp`. This means that we lose type-safety for expressions as soon as they are embedded in `$project` and `$group` ops, but parameterizing `WorkflowOpCoreF` to track the expression type became a quagmire. Currently, the expression type is tracked as far as the planner, which itself only ever generates "core" ops. Only the `FuncHandler`, which is constrained to the right type, ever generates anything fancier.